+++ src/StardustSandbox/Constants/TemperatureConstants.cs
@@ -9,39 +9,39 @@ namespace StardustSandbox.Constants
     internal static class TemperatureConstants
     {
         // Constants for simulation (can be tuned for realism)
-        internal const double THERMAL_CONDUCTIVITY = 0.5; // k, arbitrary unit
-        internal const double AREA = 1.0; // contact area, arbitrary unit
-        internal const double DISTANCE = 1.0; // distance between centers, arbitrary unit
+        internal const float THERMAL_CONDUCTIVITY = 0.5f; // k, arbitrary unit
+        internal const float AREA = 1.0f; // contact area, arbitrary unit
+        internal const float DISTANCE = 1.0f; // distance between centers, arbitrary unit
 
-        internal const double MIN_CELSIUS_VALUE = -275.0;
-        internal const double MAX_CELSIUS_VALUE = 9725.0;
-        internal const double EQUILIBRIUM_THRESHOLD = 1.0;
-        internal const double COLOR_HEAT_FACTOR = 0.45;
+        internal const float MIN_CELSIUS_VALUE = -275.0f;
+        internal const float MAX_CELSIUS_VALUE = 9725.0f;
+        internal const float EQUILIBRIUM_THRESHOLD = 1.0f;
+        internal const float COLOR_HEAT_FACTOR = 0.45f;
 
         internal static readonly TemperatureColorRange[] TEMPERATURE_COLOR_RANGES =
         [
-            new(-275.0, -200.0, AAP64ColorPalette.DarkTeal),
-            new(-199.0, -150.0, AAP64ColorPalette.NavyBlue),
-            new(-149.0, -100.0, AAP64ColorPalette.RoyalBlue),
-            new(-99.0, -50.0, AAP64ColorPalette.SkyBlue),
-            new(-49.0, 0.0, AAP64ColorPalette.Cyan),
-            new(1.0, 10.0, AAP64ColorPalette.Mint),
-            new(11.0, 20.0, AAP64ColorPalette.White),
-            new(21.0, 35.0, AAP64ColorPalette.PaleYellow),
-            new(36.0, 50.0, AAP64ColorPalette.Gold),
-            new(51.0, 100.0, AAP64ColorPalette.Orange),
-            new(101.0, 200.0, AAP64ColorPalette.OrangeRed),
-            new(201.0, 300.0, AAP64ColorPalette.Crimson),
-            new(301.0, 500.0, AAP64ColorPalette.DarkRed),
-            new(501.0, 1000.0, AAP64ColorPalette.Maroon),
-            new(1001.0, 2000.0, AAP64ColorPalette.Brown),
-            new(2001.0, 4000.0, AAP64ColorPalette.DarkBrown),
-            new(4001.0, 9725.0, AAP64ColorPalette.DarkGray)
+            new(-275.0f, -200.0f, AAP64ColorPalette.DarkTeal),
+            new(-199.0f, -150.0f, AAP64ColorPalette.NavyBlue),
+            new(-149.0f, -100.0f, AAP64ColorPalette.RoyalBlue),
+            new(-99.0f, -50.0f, AAP64ColorPalette.SkyBlue),
+            new(-49.0f, 0.0f, AAP64ColorPalette.Cyan),
+            new(1.0f, 10.0f, AAP64ColorPalette.Mint),
+            new(11.0f, 20.0f, AAP64ColorPalette.White),
+            new(21.0f, 35.0f, AAP64ColorPalette.PaleYellow),
+            new(36.0f, 50.0f, AAP64ColorPalette.Gold),
+            new(51.0f, 100.0f, AAP64ColorPalette.Orange),
+            new(101.0f, 200.0f, AAP64ColorPalette.OrangeRed),
+            new(201.0f, 300.0f, AAP64ColorPalette.Crimson),
+            new(301.0f, 500.0f, AAP64ColorPalette.DarkRed),
+            new(501.0f, 1000.0f, AAP64ColorPalette.Maroon),
+            new(1001.0f, 2000.0f, AAP64ColorPalette.Brown),
+            new(2001.0f, 4000.0f, AAP64ColorPalette.DarkBrown),
+            new(4001.0f, 9725.0f, AAP64ColorPalette.DarkGray)
         ];
 
-        internal static Color GetTemperatureColor(double temperature)
+        internal static Color GetTemperatureColor(float temperature)
         {
-            for (byte i = 0; i < TEMPERATURE_COLOR_RANGES.Length; i++)
+            for (int i = 0; i < TEMPERATURE_COLOR_RANGES.Length; i++)
             {
                 if (temperature >= TEMPERATURE_COLOR_RANGES[i].MinimumTemperature && temperature <= TEMPERATURE_COLOR_RANGES[i].MaximumTemperature)
                 {
@@ -52,7 +52,7 @@ namespace StardustSandbox.Constants
             return AAP64ColorPalette.DarkGray;
         }
 
-        internal static Color ApplyHeatColor(Color baseColor, double temperature)
+        internal static Color ApplyHeatColor(Color baseColor, float temperature)
         {
             return baseColor.OverlayBlend(GetTemperatureColor(temperature), COLOR_HEAT_FACTOR);
         }
