+++ src/StardustSandbox/Elements/Solids/Immovables/Glass.cs
@@ -4,9 +4,9 @@ namespace StardustSandbox.Elements.Solids.Immovables
 {
     internal sealed class Glass : ImmovableSolid
     {
-        protected override void OnTemperatureChanged(in ElementContext context, double currentValue)
+        protected override void OnTemperatureChanged(in ElementContext context, float currentValue)
         {
-            if (currentValue >= 620)
+            if (currentValue >= 620.0f)
             {
                 context.ReplaceElement(ElementIndex.Lava);
                 context.SetStoredElement(ElementIndex.Glass);
