+++ src/StardustSandbox/Elements/Solids/Immovables/Gold.cs
@@ -4,9 +4,9 @@ namespace StardustSandbox.Elements.Solids.Immovables
 {
     internal sealed class Gold : ImmovableSolid
     {
-        protected override void OnTemperatureChanged(in ElementContext context, double currentValue)
+        protected override void OnTemperatureChanged(in ElementContext context, float currentValue)
         {
-            if (currentValue > 1060)
+            if (currentValue > 1060.0f)
             {
                 context.ReplaceElement(ElementIndex.Lava);
                 context.SetStoredElement(ElementIndex.Gold);
