+++ src/StardustSandbox/Elements/Solids/Immovables/TreeLeaf.cs
@@ -5,9 +5,9 @@ namespace StardustSandbox.Elements.Solids.Immovables
 {
     internal sealed class TreeLeaf : ImmovableSolid
     {
-        protected override void OnTemperatureChanged(in ElementContext context, double currentValue)
+        protected override void OnTemperatureChanged(in ElementContext context, float currentValue)
         {
-            if (currentValue >= 220)
+            if (currentValue >= 220.0f)
             {
                 if (SSRandom.Chance(75))
                 {
