+++ src/StardustSandbox/Mathematics/TemperatureMath.cs
@@ -4,9 +4,9 @@ namespace StardustSandbox.Mathematics
 {
     internal static class TemperatureMath
     {
-        internal static double Clamp(double value)
+        internal static float Clamp(float value)
         {
-            return double.Clamp(value, TemperatureConstants.MIN_CELSIUS_VALUE, TemperatureConstants.MAX_CELSIUS_VALUE);
+            return float.Clamp(value, TemperatureConstants.MIN_CELSIUS_VALUE, TemperatureConstants.MAX_CELSIUS_VALUE);
         }
     }
 }
